requirements.txt
musicbrainzngs
mutagen
pycdio>0.20
PyGObject
ruamel.yaml
setuptools_scm
discid
flask
flask-table
pyudev
